% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_condition.R
\name{boundary_condition}
\alias{boundary_condition}
\title{Create boundary condition functions}
\usage{
boundary_condition(left = 0, right = 0)
}
\arguments{
\item{left}{numeric or function Left boundary ; numeric value or function of time.}

\item{right}{numeric or function Right boundary ; numeric value or function of time.}
}
\value{
description A list with two functions; left(t)and right(t).
}
\description{
Create boundary condition functions
}
\examples{
boundary_condition(left=1)
}
