% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_solution_csv.R
\name{export_solution_csv}
\alias{export_solution_csv}
\title{Export the solution matrix to a csv file}
\usage{
export_solution_csv(U, file = "solution.csv")
}
\arguments{
\item{U}{matrix Numerical solution matrix of the PDE}

\item{file}{character Name of the output CSV file (default: "solution.csv")}
}
\value{
NULL (side effect:writes a CSV file to disk)
}
\description{
Export the solution matrix to a csv file
}
\examples{
U <- matrix(runif(100), nrow= 10)
export_solution_csv(U, "solution.csv")
}
