% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_properties.R
\name{get_food_alpha}
\alias{get_food_alpha}
\title{Get Thermal Diffusivity of Food Materials}
\usage{
get_food_alpha(food_name)
}
\arguments{
\item{food_name}{character Name of the material ("water", "meat", "vegetable", "stainless_steel").}
}
\value{
numeric The thermal diffusivity value in mm^2/s.
}
\description{
Provides the thermal diffusivity (alpha) for various food items and materials
used in pasteurization simulations.
}
\examples{
alpha_val <- get_food_alpha("meat")
}
