% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_condition.R
\name{initial_condition}
\alias{initial_condition}
\title{Create initial temperature profile}
\usage{
initial_condition(x, type = "spike", value = 1, position = 0.5, sigma = 0.05)
}
\arguments{
\item{x}{numeric Spatial grid vector}

\item{type}{character Type:"constant","spike","gaussian".}

\item{value}{numeric Peak or constant value (default 1)}

\item{position}{numeric Position in 0,1 fot spike/gaussian(relative).}

\item{sigma}{numeric Standard deviation for gaussian (relatuve units).}
}
\value{
Numeric vector of same length as x (initial temperature).
}
\description{
Create initial temperature profile
}
\examples{
initial_condition(c(1,55,8))
}
