% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_grid.R
\name{initialize_grid}
\alias{initialize_grid}
\title{Initialize spatial and temporal grid}
\usage{
initialize_grid(L = 1, nx = 101, t = 1, nt = NULL)
}
\arguments{
\item{L}{numeric Length of the spatial domain.}

\item{nx}{integer Number of spatial grid points.}

\item{t}{Total simulation time.}

\item{nt}{integer Number of time steps.}
}
\value{
A list with elements: x (space vector)
}
\description{
Initialize spatial and temporal grid
}
\examples{
grid <- initialize_grid()
}
