% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_heat1d_cn.R
\name{solve_heat1d_cn}
\alias{solve_heat1d_cn}
\title{Solve the one-dimensional heat equation using the Crank-Nicolson method}
\usage{
solve_heat1d_cn(u0, L, alpha, dt, t, bc = list(left = 0, right = 0))
}
\arguments{
\item{u0}{numeric vector Initial temperature}

\item{L}{length of the bar}

\item{alpha}{thermal diffusivity}

\item{dt}{time step}

\item{t}{total time}

\item{bc}{list (left=0, right=0)}
}
\value{
Matrix solution
}
\description{
Solve the one-dimensional heat equation using the Crank-Nicolson method
}
\examples{
u0 <- rep(20, 21) # Vecteur de temperature initiale
U <- solve_heat1d_cn(u0, L=1, alpha=0.1, dt=0.001, t=0.1, bc=list(left=80, right=80))
}
