% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_heat1d_explicit.R
\name{solve_heat1d_explicit}
\alias{solve_heat1d_explicit}
\title{Solve Heat Equation(Explicit method)}
\usage{
solve_heat1d_explicit(u0, L, alpha, dt, t, bc = list(left = 0, right = 0))
}
\arguments{
\item{u0}{initial temperature vector}

\item{L}{length of the bar}

\item{alpha}{thermal diffusivity}

\item{dt}{time step}

\item{t}{final time}

\item{bc}{list (boundary_left=0, boundary_right=0)}
}
\value{
matrix of solution
}
\description{
Solve Heat Equation(Explicit method)
}
\examples{
u0 <- sin(seq(0, pi, length.out=20))
U <- solve_heat1d_explicit(u0, L=1, alpha=0.1, dt=0.001, t=0.1)
}
