% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thomas_solve.R
\name{thomas_solve}
\alias{thomas_solve}
\title{Algorithme de Thomas pour systeme tri-diagonal}
\usage{
thomas_solve(a, b, c, d)
}
\arguments{
\item{a}{sous-diagonale}

\item{b}{diagonale principale}

\item{c}{sur-diagonale}

\item{d}{vecteur membre de droite}
}
\description{
Algorithme de Thomas pour systeme tri-diagonal
}
